+++ src/data/constant/InvokeDynamicConstant.java
@@ -8,24 +8,24 @@ import static util.Util.writeShort;
 
 public class InvokeDynamicConstant implements ConstantDesc {
 
-    private final short bootstrapMethodAttributeIndex;
-    private final short nameAndTypeIndex;
+    private final int bootstrapMethodAttributeIndex;
+    private final int nameAndTypeIndex;
 
-    public InvokeDynamicConstant(short bootstrapMethodAttributeIndex, short nameAndTypeIndex) {
+    public InvokeDynamicConstant(int bootstrapMethodAttributeIndex, int nameAndTypeIndex) {
         this.bootstrapMethodAttributeIndex = bootstrapMethodAttributeIndex;
         this.nameAndTypeIndex = nameAndTypeIndex;
     }
 
     @Override
-    public byte getTag() {
+    public int getTag() {
         return 18;
     }
 
-    public short getBootstrapMethodAttributeIndex() {
+    public int getBootstrapMethodAttributeIndex() {
         return bootstrapMethodAttributeIndex;
     }
 
-    public short getNameAndTypeIndex() {
+    public int getNameAndTypeIndex() {
         return nameAndTypeIndex;
     }
