+++ src/data/constant/LongConstant.java
@@ -2,8 +2,8 @@ package data.constant;
 
 import data.ClassFile;
 import data.ConstantDesc;
-import java.io.IOException;
 import java.io.DataInputStream;
+import java.io.IOException;
 import java.io.OutputStream;
 import static util.Util.writeLong;
 
@@ -16,7 +16,7 @@ public class LongConstant implements ConstantDesc {
     }
 
     @Override
-    public byte getTag() {
+    public int getTag() {
         return 5;
     }
